requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.6
django-apscheduler==0.6.0
django-heroku==0.0.0
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.2
protobuf==3.17.3
python-dateutil==2.8.2
python-string-utils==1.0.0
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
whitenoise==5.3.0